requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ecdsa==0.19.0
fastapi==0.115.9
frozendict==2.4.6
h11==0.14.0
idna==3.10
multitasking==0.0.11
numpy==2.2.3
pandas==2.2.3
passlib==1.7.4
peewee==3.17.9
platformdirs==4.3.6
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
uvicorn==0.34.0
yfinance==0.2.54
